[TRAFODION-2700] Avoid parallel plans for single salt partitions
When costing an HBase scan, limit the effective degree of parallelism
to the minimum of the number of partitions (regions) read and the
number of ESPs.

The effective DoP for both the region server and client sides is
the minimum of the two: If there are fewer regions than clients
(ESPs), then some ESPs have nothing to do. If there are more
regions than clients, the clients will process multiple regions
sequentially, limiting the DoP on the region servers.